pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "dynogrid"
version = "0.1.0"
description = "Dynamic grid algorithmic trading bot with paper trading first."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "aiosqlite>=0.20",
  "ccxt>=4.0",
  "pandas>=2.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[project.scripts]
dynogrid = "dynogrid.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
